do not install yq if its already there

+++ install/tools/terminal.sh
@@ -6,5 +6,7 @@ sudo apt install -y \
   wl-clipboard htop \
   man tldr less whois plocate bash-completion
 
-sudo wget https://github.com/mikefarah/yq/releases/latest/download/yq_linux_amd64 -O /usr/local/bin/yq
-sudo chmod +x /usr/local/bin/yq
+if [ ! -f '/usr/local/bin/yq' ]; then
+  sudo wget https://github.com/mikefarah/yq/releases/latest/download/yq_linux_amd64 -O /usr/local/bin/yq
+  sudo chmod +x /usr/local/bin/yq
+fi
